Rename pkg comp BaseMap to GenericBaseMap
