--- src/components/BaseMap/package.json
@@ -1,6 +1,0 @@
-{
-  "name": "BaseMap",
-  "version": "0.0.0",
-  "private": true,
-  "main": "./BaseMap.js"
-}
