+++ src/index.js
@@ -1,3 +1,3 @@
 export { default as HelloWorld } from './components/HelloWorld';
-export { default as BaseMap } from './components/BaseMap';
+export { default as GenericBaseMap } from './components/GenericBaseMap';
 export { default as BCBaseMap } from './components/BCBaseMap';
